修复查看自评参数类型异常

+++ ruoyi-teacher/src/main/java/com/ruoyi/teacher/domain/vo/CompletionRateRange.java
@@ -12,6 +12,6 @@ public class CompletionRateRange {
     private String range;
 
     /**人数*/
-    private Long peopleNum;
+    private Integer peopleNum;
 
 }
